refactor(web): repoint remaining AppRouter callers at server/api/router
Address PR #522 review feedback. The Phase 0 scaffold left several
callers still importing AppRouter / the OpenAPI source path from the
legacy `src/trpc/router.ts` location. Update them all to the new
canonical entry point at `src/server/api/router.ts` so future phases
that move procedures out of the monolithic file don't silently break
these consumers:

- apps/web/start-server.ts — dev-mode `generateOpenAPIDocument` source
- apps/web/scripts/build-server.sh — prod-build `trpc-openapi` CLI input
- apps/web/src/mcp/server.ts — runtime `appRouter` import for MCP bridge
- apps/web/src/lib/trpc-client.ts — `AppRouter` type for the web client
- apps/web/tests/trpc-batch-url.test.ts — `AppRouter` type for batch test
- apps/web/e2e/helpers/trpc-mock.ts — `AppRouter` type for e2e mock layer

No behavioural change: `server/api/router.ts` is still a re-export of
`src/trpc/router.ts` in Phase 0, so this is purely path consistency.
