+++ apps/web/scripts/build-server.sh
@@ -3,7 +3,7 @@ set -euo pipefail
 
 # Generate OpenAPI spec from tRPC router (static TypeScript analysis)
 mkdir -p dist
-pnpm exec trpc-openapi ./src/trpc/router.ts -o dist/openapi.json --title "Band API" --version "1.0.0"
+pnpm exec trpc-openapi ./src/server/api/router.ts -o dist/openapi.json --title "Band API" --version "1.0.0"
 
 # Bundle the server entry point
 # `vite` and `@trpc/openapi` are imported dynamically inside the
