+++ apps/web/tests/trpc-batch-url.test.ts
@@ -34,7 +34,7 @@ import { join } from "node:path";
 import { createTRPCClient, httpBatchLink } from "@trpc/client";
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
 import { toWorkspaceId } from "@/dashboard";
-import type { AppRouter } from "../src/trpc/router";
+import type { AppRouter } from "../src/server/api/router";
 import { seedSettings, seedState } from "./helpers/seed-state";
 import { SERVER_RUNTIME, SERVER_SCRIPT } from "./helpers/server-runtime";
